added Lamp and 3/4s Lamp screenshot.
I think it looks pretty good. I have about 50 min to make the 3rd object, so I am a little short on time so I wasn't as intricate inside the hood as I wanted to be. But the lightbulb came out with a fine shape and I did it in like 30 seconds, including the restart, so I'm happy enough.